scandir header
* moved utility function scandir(path) to its own header (to be included
* in more executables)
