+++ src/rm.cpp
@@ -12,6 +12,7 @@
 
 
 #include "rm.h"
+#include "scandir.h"
 
 
 bool RM_DIRS = false;
@@ -80,44 +81,6 @@ int removedir(const char* path) {
 
     // tail recursion; delete current dir
     if(rmdir(path) != 0) { perror("rmdir"); return 1; }
-
-    return 0;
-}
-
-
-
-// yanked from previous project
-// scan a directory for all contained files
-// return values: -1 - syscall error; 1 - is a file
-int scandir(const char* path, std::vector<std::string> &files) {
-   
-    struct stat stat_buf;
-    if(stat(path, &stat_buf) == -1) { perror("stat"); return 1; }
-
-    if(!S_ISDIR(stat_buf.st_mode))
-        return 1;
-
-	DIR* dirp;
-    if( (dirp = opendir(path)) == NULL) {
-        perror("opendir");
-        return -1; 
-    }
-
-
-    struct dirent* entry;
-    errno = 0;
-    while( (entry = readdir(dirp)) != NULL)
-        files.push_back(entry->d_name);
-    
-    if(errno != 0) {
-        perror("readdir");
-        return -1;
-    }
-
-   if(closedir(dirp) != 0) {
-        perror("closedir");
-        return -1;
-    }
 
     return 0;
 }
