+++ src/rm.h
@@ -3,7 +3,6 @@
 #define RM_H
 
 int removedir(const char*);
-int scandir(const char*, std::vector<std::string> &);
 
 #endif
